fixed typecasting error

+++ src/utils/perms.rs
@@ -3,9 +3,9 @@ use std::os::unix::fs::{PermissionsExt};
 
 pub fn perms(file: std::path::PathBuf) -> String {
   let mode = file.symlink_metadata().unwrap().permissions().mode() as u16;
-  let user = masking(mode, S_IRUSR as u16, S_IWUSR as u16, S_IXUSR);
-  let group = masking(mode, S_IRGRP as u16, S_IWGRP as u16, S_IXGRP);
-  let other = masking(mode, S_IROTH as u16, S_IWOTH as u16, S_IXOTH);
+  let user = masking(mode, S_IRUSR as u16, S_IWUSR as u16, S_IXUSR as u16);
+  let group = masking(mode, S_IRGRP as u16, S_IWGRP as u16, S_IXGRP as u16);
+  let other = masking(mode, S_IROTH as u16, S_IWOTH as u16, S_IXOTH as u16);
   let f = crate::PathType::new(&file).unwrap()[0].get_letter_for_type();
   [f, user, group, other].join("")
 }
